restore setup.py
Spent too many hours trying to make this work, and pyproject.toml only
generated broken packages.

I can't find a way to publish rxdjango with setup.py either.
For some reason, cannot properly authenticate. This used to work,
I also spent many hours on this.

For now, I'm giving up releasing rxdjango, as this has became the
biggest bottleneck in project development. With setup.py I can't publish,
but at least I generate functional packages for our project to link from
repository directly.

It's so frustrating to have a functional code and not to be able to release,
but it's been too much time spent on this, and too long having development
blocked because we made this a priority.

Moving on, not releasing for now.