Fix synchronization bug in heating cache state
The bug was caused by an initial subscribe message to the pubsub
channel. While debugging the synchronization, fixed some missing
links.

Also changed RedisStateSession.write_instances to do several operations
instead of a Lua script. This operation does not need to be atomic,
so it's better to not hold redis with a big atomic operation.