Optimize delta calculations in C
Every time an instance is saved, the new version is serialized
and it's compared with the current version in cache, so only deltas
are sent to clients. This commit separates this routine in the
rxdjango.utils.delta_utils module and creates a rxdjango.utils.delta_utils_c
module which has the same interface and implemented in C.

The mongo code will try to get the C implementation first, then use the python
one if C is not available.

Profiling shows this is 1.5 to 2.7 times faster